backend/requirements.txt
flask==2.3.3
python-dotenv==1.0.1
gunicorn==23.0.0
flask-cors==5.0.0
waitress==2.1.2
pymssql==2.2.7
